Cache settings to avoid going to the DB everytime
