+++ Rocket.Chat/Managers/Socket/SocketManager.swift
@@ -112,7 +112,7 @@ extension SocketManager {
             }
 
             SubscriptionManager.updateSubscriptions(auth, completion: { _ in
-                AuthManager.updatePublicSettings(auth, completion: { _ in
+                AuthSettingsManager.updatePublicSettings(auth, completion: { _ in
 
                 })
